propelize-master/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==25.2.0
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
uritemplate==4.1.1
